requirements.txt
Pillow
requests
transformers
torch
pydantic
python-dotenv
